remove slot in audit header
